requirements.txt
# Google ADK — install via: pip install google-adk
# (package name may vary; check https://google.github.io/adk-docs/)
google-adk

# HTTP client used by football-data.org integration
requests>=2.31.0

# Google Maps — geocoding + Places Nearby (requires GOOGLE_MAPS_API_KEY)
googlemaps>=4.10.0

# Load .env variables into the process environment
python-dotenv>=1.0.0
